fix: rename command directory
